fix(34): verification fixes — u64 BigInt, deserialization, built-in lifecycle
- Change RateLimitInfo.reset from u64 to u32 (specta forbids BigInt)
- Add internal GitHubDeviceCodeResponse for snake_case GitHub API parsing
- Set real GitHub OAuth client_id
- Add builtIn flag to ExtensionInfo to survive deactivateAll/discoverExtensions
- Preserve built-in extensions in discoverExtensions Map rebuild
- Skip built-in extensions in deactivateAll

+++ src-tauri/src/github/rate_limit.rs
@@ -22,7 +22,7 @@ struct RateLimitResources {
 struct RateLimitData {
     limit: u32,
     remaining: u32,
-    reset: u64,
+    reset: u32,
     used: u32,
 }
